Fix: User was always redirected to home page
Closes: #27
